Revert changes to original harness, use new harness
This commit effectively reverts the following commits, and reapplies them to a
new file called `bat_stdin_input_random_options`:
- 25fd500
- 664ded8
- cc72450
- b1189d2
